+++ core/src/main/java/org/musetest/core/context/TestExecutionContext.java
@@ -7,7 +7,6 @@ import org.musetest.core.*;
  */
 public interface TestExecutionContext extends MuseExecutionContext
     {
-//    TestResultProducer getTestResultSource();  // TODO
     MuseTest getTest();
     void addInitializer(ContextInitializer initializer);
     }