+++ core/src/test/java/org/musetest/core/tests/TestResultProducerTests.java
@@ -38,7 +38,7 @@ public class TestResultProducerTests
         TestResultProducer producer = new TestFailsOnErrorFailureOrInterrupt(test, log);
 
         String reason = "test123";
-        producer.eventRaised(new StepEvent(MuseEventType.EndStep, new StepConfiguration(LogMessage.TYPE_ID), new DummyStepExecutionContext(new SimpleProject()), new BasicStepExecutionResult(StepExecutionStatus.FAILURE, reason)));
+        producer.eventRaised(new StepEvent(MuseEventType.EndStep, new StepConfiguration(LogMessage.TYPE_ID), new MockStepExecutionContext(new SimpleProject()), new BasicStepExecutionResult(StepExecutionStatus.FAILURE, reason)));
 
         MuseTestResult result = producer.getTestResult();
         Assert.assertFalse(result.isPass());
@@ -55,7 +55,7 @@ public class TestResultProducerTests
         TestResultProducer producer = new TestFailsOnErrorFailureOrInterrupt(test, log);
 
         String reason = "test123";
-        producer.eventRaised(new StepEvent(MuseEventType.EndStep, new StepConfiguration(LogMessage.TYPE_ID), new DummyStepExecutionContext(new SimpleProject()), new BasicStepExecutionResult(StepExecutionStatus.ERROR, reason)));
+        producer.eventRaised(new StepEvent(MuseEventType.EndStep, new StepConfiguration(LogMessage.TYPE_ID), new MockStepExecutionContext(new SimpleProject()), new BasicStepExecutionResult(StepExecutionStatus.ERROR, reason)));
 
         MuseTestResult result = producer.getTestResult();
         Assert.assertFalse(result.isPass());
@@ -72,7 +72,7 @@ public class TestResultProducerTests
         TestResultProducer producer = new TestFailsOnErrorFailureOrInterrupt(test, log);
 
         String reason = "verifyfail";
-        producer.eventRaised(new VerifyFailureEvent(new StepConfiguration(LogMessage.TYPE_ID), new DummyStepExecutionContext(new SimpleProject()), reason));
+        producer.eventRaised(new VerifyFailureEvent(new StepConfiguration(LogMessage.TYPE_ID), new MockStepExecutionContext(new SimpleProject()), reason));
 
         MuseTestResult result = producer.getTestResult();
         Assert.assertFalse(result.isPass());