+++ change-log.md
@@ -4,6 +4,7 @@
 - add: RandomNumberValueSource and random(min,max) expression support.
 - fix: Implement createResource() for Macro
 - fix: Ensure .json and .csv files are closed after reading by the responsible resource factories.
+- new: Verify step adds (optional) parameter to terminate test on failure
 
 ## 0.7
 - add: Change listeners on WebDriverProviderList